+++ srcs/traverse/sh_traverse_cmd_name.c
@@ -3,16 +3,16 @@
 /*                                                        :::      ::::::::   */
 /*   sh_traverse_cmd_name.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jmartel <jmartel@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ldedier <ldedier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/15 17:31:30 by ldedier           #+#    #+#             */
-/*   Updated: 2019/09/23 14:51:43 by jmartel          ###   ########.fr       */
+/*   Updated: 2019/11/16 09:05:16 by ldedier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "sh_21.h"
 
-int			sh_traverse_cmd_name(t_ast_node *node, t_context *context)
+long		sh_traverse_cmd_name(t_ast_node *node, t_context *context)
 {
 	return (sh_traverse_cmd_suffix(node, context));
 }
