+++ srcs/traverse/sh_traverse_cmd_suffix.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sh_traverse_cmd_suffix.c                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jmartel <jmartel@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ldedier <ldedier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/15 17:31:30 by ldedier           #+#    #+#             */
-/*   Updated: 2019/11/04 12:26:42 by jdugoudr         ###   ########.fr       */
+/*   Updated: 2019/11/16 09:05:16 by ldedier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ static int	sh_process_traverse_cmd_suffix(
 	return (SUCCESS);
 }
 
-int			sh_traverse_cmd_suffix(t_ast_node *node, t_context *context)
+long		sh_traverse_cmd_suffix(t_ast_node *node, t_context *context)
 {
 	t_ast_node	*child;
 	t_list		*ptr;
