+++ srcs/traverse/sh_traverse_simple_command.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sh_traverse_simple_command.c                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mdaoud <mdaoud@student.42.fr>              +#+  +:+       +#+        */
+/*   By: ldedier <ldedier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/15 17:34:52 by ldedier           #+#    #+#             */
-/*   Updated: 2019/11/09 00:51:19 by mdaoud           ###   ########.fr       */
+/*   Updated: 2019/11/16 09:05:16 by ldedier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ static int	get_cmd_string(t_ast_node *node, t_context *context)
 **	any value returned by a builtin executed or a process launched
 */
 
-int			sh_traverse_simple_command(t_ast_node *node, t_context *context)
+long		sh_traverse_simple_command(t_ast_node *node, t_context *context)
 {
 	int		ret;
