+++ srcs/traverse/sh_traverse_subshell.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   sh_traverse_subshell.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mdaoud <mdaoud@student.42.fr>              +#+  +:+       +#+        */
+/*   By: ldedier <ldedier@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/02 10:03:30 by jdugoudr          #+#    #+#             */
-/*   Updated: 2019/11/08 23:37:38 by mdaoud           ###   ########.fr       */
+/*   Updated: 2019/11/16 09:05:16 by ldedier          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,7 +73,7 @@ static int	parents_part(pid_t pid, t_context *context)
 ** So we can directly go to compound_list node.
 */
 
-int			sh_traverse_subshell(t_ast_node *node, t_context *ctxt)
+long		sh_traverse_subshell(t_ast_node *node, t_context *ctxt)
 {
 	pid_t	pid;
 	int		ret;
